9593: move helper function
