+++ shared/src/business/dto/docketEntries/PublicCaseDTO.ts
@@ -2,7 +2,7 @@ import { RawPublicContact } from '@shared/business/entities/cases/PublicContact'
 import { IrsPractitioner } from '@shared/business/entities/IrsPractitioner';
 import { PrivatePractitioner } from '@shared/business/entities/PrivatePractitioner';
 import { ConsolidatedCaseSummary } from '../cases/ConsolidatedCaseSummary';
-import { removeServedParties } from './CaseDTO';
+import { removeServedParties } from '../helpers/removeServedParties';
 
 export class PublicCaseDTO {
   public entityName: string;
